+++ Sources/Core/Classes/models/requests/SignupRequest.swift
@@ -4,10 +4,10 @@ public class SignupRequest: Codable, DictionaryEncodable {
     public let clientId: String
     public let data: ProfileSignupRequest
     public let scope: String
-    public let redirectUrl: String?
+    public let redirectUrl: URL?
     public let origin: String?
-    
-    public init(clientId: String, data: ProfileSignupRequest, scope: String, redirectUrl: String?, origin: String? = nil) {
+
+    public init(clientId: String, data: ProfileSignupRequest, scope: String, redirectUrl: URL?, origin: String? = nil) {
         self.clientId = clientId
         self.data = data
         self.scope = scope