+++ Sources/Core/Classes/models/requests/VerifyEmailRequest.swift
@@ -1,9 +1,9 @@
 import Foundation
 
 public class SendEmailVerificationRequest: Codable, DictionaryEncodable {
-    public let redirectUrl: String?
-    
-    public init(redirectUrl: String? = nil) {
+    public let redirectUrl: URL?
+
+    public init(redirectUrl: URL? = nil) {
         self.redirectUrl = redirectUrl
     }
 }
@@ -11,10 +11,10 @@ public class SendEmailVerificationRequest: Codable, DictionaryEncodable {
 public class VerifyEmailRequest: Codable, DictionaryEncodable {
     public let email: String
     public let verificationCode: String
-    
+
     public init(email: String, verificationCode: String) {
         self.email = email
         self.verificationCode = verificationCode
     }
-    
+
 }